Added TournamentCollection, CallGroup and Options endpoints.
Please keep the end goal outlined in the following PR in mind as you
consider accepting the PR

https://github.com/ccpgames/eveonline-third-party-documentation/pull/212#issuecomment-244253260

A new documentation pattern has been suggested which implies all
endpoints would adopt the pattern.  See the TournamentCollection for
temporary comments about the pattern.

I have not completed the new ###Data section on the CallGroup or Options
endpoints.  I will complete these upon approval of the proposed pattern.
I also suck at md and not sure of a nice layout to represent an objects
typing yet.

The TournamentCollection (short the temporary comments) is an example of
a place holder for endpoints discovered via the schema but not yet
documented.