misc index .. does link need a "./" ?
TournamentGetTournament and TournamentGetSeries Data sections filled out

Per these links returned object graphs: 

https://api-sisi.testeveonline.com/tournaments/9/ 

https://api-sisi.testeveonline.com/tournaments/9/series/ 


I have the following still to investigate:

TournamentMatchCollection
https://api-sisi.testeveonline.com/tournaments/9/series/0/matches/

TournamentGetFrame
https://api-sisi.testeveonline.com/tournaments/9/series/0/matches/0/realtime/0/

And these are only for "Route" endpoints. Several of the Data's themself
have urls to other data structures yet to be investigated.

Because of all the files, I went ahead and created a "Misc" group and
moved all the Tournament .md's into a .../misc/tournament folder.

Some response from cpp would be nice to know if this is all a waste of
time.